try/requirements.txt
pandas
numpy==1.26.4
openai==1.56.1
requests==2.32.3

langchain==0.3.9
langchain-core==0.3.25
langchain-openai==0.2.12
langchainhub==0.1.15

python-dotenv==1.0.1
tqdm==4.66.4
pyyaml==6.0.1
tenacity==9.1.2

python-docx